[refactor] restructering folders
