+++ TelemetryDevice/Controllers/TelemetryDeviceController.cs
@@ -4,8 +4,8 @@ using Microsoft.AspNetCore.Mvc;
 using TelemetryDevices.Common;
 using TelemetryDevices.Dto;
 using TelemetryDevices.Models;
-using TelemetryDevices.Services;
 using TelemetryDevices.Services.PortsManager;
+using TelemetryDevices.Services.TelemetryDevicesManager;
 
 namespace TelemetryDevices.Controllers
 {
