+++ TelemetryDevice/Services/PipeLines/Blocks/Decoder/North_South/TelemetryDecoderBlock.cs
@@ -2,9 +2,9 @@
 using System.Threading.Tasks.Dataflow;
 using Core.Common.Enums;
 using Core.Models.ICDModels;
+using TelemetryDevices.Extensions;
+using TelemetryDevices.Helpers;
 using TelemetryDevices.Models;
-using TelemetryDevices.Services.Extensions;
-using TelemetryDevices.Services.Helpers;
 
 namespace TelemetryDevices.Services.PipeLines.Blocks.Decoder.North_South
 {
