+++ TelemetryDevice/TelemetryDevices.csproj
@@ -24,4 +24,7 @@
   <ItemGroup>
     <ProjectReference Include="..\..\Core\Core\Core.csproj" />
   </ItemGroup>
+  <ItemGroup>
+    <Folder Include="Services\Quartz\Jobs\" />
+  </ItemGroup>
 </Project>
